refactor: migrate form components to shared structure
- Updated form components to use a unified structure under '@shared/components/form'.
- Removed legacy UI form components to streamline the codebase.
- Adjusted imports in FoodForm, form-view, and Storybook stories to reflect the new component paths.
- Enhanced Storybook examples to demonstrate the updated form components.
